requirements.txt
flask==1.1.2
Pillow==7.1.2